refactor : activity folder structure + responsive modals
