+++ src/components/student-profile/activity/add-activity-modal.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import React, { useState } from 'react'
+import React from 'react'
 
 import { useToast } from '@/components/ui/use-toast'
 
@@ -19,17 +19,15 @@ import {
 } from '../../ui/select'
 import {
   Dialog,
-  DialogClose,
   DialogContent,
   DialogFooter,
   DialogTrigger,
+  DialogClose,
 } from '@/components/ui/dialog'
 
-import { FilePenLine } from 'lucide-react'
+import { Plus } from 'lucide-react'
 
-import { editActivity } from '@/lib/functions/http/edit-activity-req'
-
-import { Activity } from '../../../../types/types'
+import { addActivity } from '@/lib/functions/http/add-activity-req'
 
 const ACTIVITY_TYPES = [
   { code: 'RESEARCH', name: 'Pesquisa' },
@@ -38,56 +36,34 @@ const ACTIVITY_TYPES = [
   { code: 'INTERNSHIP', name: 'Estágio' },
 ]
 
-const formatDate = (dateString: string) => {
-  const [day, month, year] = dateString.split('/')
-  return `${year}-${month}-${day}`
-}
-
-const EditActivityModal = ({
-  title,
-  type,
-  description,
-  startDate,
-  endDate,
-  shift,
-  payment,
-  inProgress,
-}: Activity) => {
+const AddActivityModal = () => {
   const { toast } = useToast()
 
-  const [open, setOpen] = useState(false)
-  const [loading, setLoading] = useState(false)
+  const [open, setOpen] = React.useState(false)
+  const [loading, setLoading] = React.useState(false)
 
-  const [activityName, setActivityName] = useState(title || '')
-  const [activityDescription, setActivityDescription] = useState(
-    description || '',
-  )
-  const [activityWorkShift, setActivityWorkShift] = useState(shift || '')
-  const [activityType, setActivityType] = useState(type.code || 'RESEARCH')
-  const [activityStartDate, setActivityStartDate] = useState(
-    formatDate(startDate) || '',
-  )
-  const [activityEndDate, setActivityEndDate] = useState(
-    formatDate(endDate) || '',
-  )
+  const [name, setName] = React.useState('')
+  const [description, setDescription] = React.useState('')
+  const [workShift, setWorkShift] = React.useState('')
+  const [type, setType] = React.useState('')
+  const [startDate, setStartDate] = React.useState('')
+  const [endDate, setEndDate] = React.useState('')
 
-  const [activityStatusInProgress, setActivityStatusInProgress] = useState(
-    inProgress || false,
-  )
-  const [activityStatusPaid, setActivityStatusPaid] = useState(payment || false)
+  const [inProgress, setInProgress] = React.useState(false)
+  const [statusPaid, setStatusPaid] = React.useState(false)
 
-  const [nameError, setNameError] = useState('')
-  const [descriptionError, setDescriptionError] = useState('')
-  const [workShiftError, setWorkShiftError] = useState('')
-  const [typeError, setTypeError] = useState('')
-  const [startDateError, setStartDateError] = useState('')
-  const [endDateError, setEndDateError] = useState('')
+  const [nameError, setNameError] = React.useState('')
+  const [descriptionError, setDescriptionError] = React.useState('')
+  const [workShiftError, setWorkShiftError] = React.useState('')
+  const [typeError, setTypeError] = React.useState('')
+  const [startDateError, setStartDateError] = React.useState('')
+  const [endDateError, setEndDateError] = React.useState('')
 
   const validate = () => {
     let isValid = true
 
-    // Validalção do nome
-    if (activityName.trim() === '') {
+    // Validação do nome
+    if (name.trim() === '') {
       setNameError('Campo obrigatório')
       isValid = false
     } else {
@@ -95,7 +71,7 @@ const EditActivityModal = ({
     }
 
     // Validação da descrição
-    if (activityDescription.trim() === '') {
+    if (description.trim() === '') {
       setDescriptionError('Campo obrigatório')
       isValid = false
     } else {
@@ -103,7 +79,7 @@ const EditActivityModal = ({
     }
 
     // Validação da dedicação semanal
-    if (activityWorkShift.trim() === '') {
+    if (workShift.trim() === '') {
       setWorkShiftError('Campo obrigatório')
       isValid = false
     } else {
@@ -111,11 +87,7 @@ const EditActivityModal = ({
     }
 
     // Validação do tipo
-    if (
-      !ACTIVITY_TYPES.map((activity) => activity.code).includes(
-        activityType.toString(),
-      )
-    ) {
+    if (!ACTIVITY_TYPES.map((activity) => activity.code).includes(type)) {
       setTypeError('Campo obrigatório')
       isValid = false
     } else {
@@ -131,37 +103,35 @@ const EditActivityModal = ({
     }
 
     // Validação da data de término
-    if (!activityStatusInProgress && endDate.trim() === '') {
+    if (!inProgress && endDate.trim() === '') {
       setEndDateError('Campo obrigatório')
       isValid = false
     } else {
       setEndDateError('')
     }
-
     return isValid
   }
-
   const handleSubmit = async () => {
     if (!validate()) {
       return
     }
 
     const data = {
-      name: activityName,
-      type: activityType,
-      description: activityDescription,
-      startDate: activityStartDate,
-      endDate: activityEndDate,
-      workShift: activityWorkShift,
-      payment: activityStatusPaid,
-      inProgress: activityStatusInProgress,
+      name,
+      description,
+      workShift,
+      type,
+      startDate,
+      endDate,
+      inProgress,
+      statusPaid,
     }
 
     console.log(data)
 
     // setLoading(true)
 
-    // const res = await editActivity(data)
+    // const res = await addActivity(data)
 
     // if (res) {
     //   await new Promise((resolve) => setTimeout(resolve, 500))
@@ -183,26 +153,22 @@ const EditActivityModal = ({
   return (
     <Dialog open={open} onOpenChange={setOpen}>
       <DialogTrigger asChild>
-        <span className="flex items-center gap-1">
-          <FilePenLine
-            size={15}
-            className="hover:cursor-pointer hover:text-muted-foreground"
-          />
-          {/* <p>Editar</p> */}
+        <span>
+          <Plus className="h-6 w-6 cursor-pointer" />
         </span>
       </DialogTrigger>
-      <DialogContent className="p-0">
+      <DialogContent className="p-0 overflow-y-scroll max-h-[95vh]">
         <div
           key="1"
           className="bg-background border p-6 rounded-lg shadow max-w-lg mx-auto"
         >
           <h3 className="text-xl font-semibold mb-4">
-            Editar uma atividade extra
+            Adicionar uma atividade extra
           </h3>
           <p className="text-sm text-foreground mb-6">
-            Edite a atividade <span className="font-bold">{activityName}</span>
-            utilizando o formulário abaixo. Evite deixar informações
-            desatualizadas.
+            Adicione aqui uma atividade extra que você realizou ou está
+            realizando. Atividades extras podem incluir Pesquisa, Monitoria,
+            Estágio, Extensão ou outras atividades similares.
           </p>
           <form>
             <div className="mb-4">
@@ -212,27 +178,25 @@ const EditActivityModal = ({
               >
                 Tipo da atividade*
               </label>
-              <Select
-                onValueChange={(value) =>
-                  setActivityType(
-                    value as
-                      | 'RESEARCH'
-                      | 'EXTENSION'
-                      | 'TUTORING'
-                      | 'INTERNSHIP',
-                  )
-                }
-                defaultValue={activityType}
-              >
+              <Select onValueChange={(value) => setType(value)}>
                 <SelectTrigger id="activity-type">
-                  <SelectValue placeholder="Selecione o tipo da atividade" />
+                  <SelectValue placeholder="Selecione a atividade" />
                 </SelectTrigger>
                 <SelectContent position="popper">
                   <SelectGroup>
-                    <SelectItem value="RESEARCH">Pesquisa</SelectItem>
-                    <SelectItem value="EXTENSION">Extensão</SelectItem>
-                    <SelectItem value="TUTORING">Monitoria</SelectItem>
-                    <SelectItem value="INTERNSHIP">Estágio</SelectItem>
+                    {ACTIVITY_TYPES.map((activity) => (
+                      <div
+                        key={activity.code}
+                        className="w-[--radix-select-trigger-width]"
+                      >
+                        <SelectItem
+                          value={activity.code}
+                          className="hover:cursor-pointer"
+                        >
+                          {activity.name}
+                        </SelectItem>
+                      </div>
+                    ))}
                   </SelectGroup>
                 </SelectContent>
               </Select>
@@ -250,8 +214,8 @@ const EditActivityModal = ({
               <Input
                 id="activity-name"
                 placeholder=""
-                value={activityName}
-                onChange={(e) => setActivityName(e.target.value)}
+                value={name}
+                onChange={(e) => setName(e.target.value)}
               />
               {nameError && (
                 <span className="text-red-500 text-sm">{nameError}</span>
@@ -268,7 +232,7 @@ const EditActivityModal = ({
                 id="description"
                 placeholder="Escreva uma descrição detalhada da atividade."
                 value={description}
-                onChange={(e) => setActivityDescription(e.target.value)}
+                onChange={(e) => setDescription(e.target.value)}
               />
               {descriptionError && (
                 <span className="text-red-500 text-sm">{descriptionError}</span>
@@ -285,8 +249,8 @@ const EditActivityModal = ({
                 id="weekly-dedication"
                 placeholder="20"
                 type="number"
-                value={activityWorkShift}
-                onChange={(e) => setActivityWorkShift(e.target.value)}
+                value={workShift}
+                onChange={(e) => setWorkShift(e.target.value)}
               />
               {workShiftError && (
                 <span className="text-red-500 text-sm">{workShiftError}</span>
@@ -304,8 +268,8 @@ const EditActivityModal = ({
                   id="start-date"
                   placeholder="10/09/2023"
                   type="date"
-                  value={activityStartDate}
-                  onChange={(e) => setActivityStartDate(e.target.value)}
+                  value={startDate}
+                  onChange={(e) => setStartDate(e.target.value)}
                 />
                 {startDateError && (
                   <span className="text-red-500 text-sm">{startDateError}</span>
@@ -322,9 +286,9 @@ const EditActivityModal = ({
                   id="end-date"
                   placeholder="01/10/2024"
                   type="date"
-                  value={activityEndDate}
-                  onChange={(e) => setActivityEndDate(e.target.value)}
-                  disabled={activityStatusInProgress}
+                  value={endDate}
+                  onChange={(e) => setEndDate(e.target.value)}
+                  disabled={inProgress}
                 />
                 {endDateError && (
                   <span className="text-red-500 text-sm">{endDateError}</span>
@@ -335,10 +299,10 @@ const EditActivityModal = ({
               <div className="flex items-center mb-2">
                 <Switch
                   id="status-in-progress"
-                  checked={activityStatusInProgress}
+                  checked={inProgress}
                   onCheckedChange={(checked) => {
-                    setActivityStatusInProgress(checked)
-                    setActivityEndDate('')
+                    setInProgress(checked)
+                    setEndDate('')
                   }}
                 />
                 <Label className="ml-2 text-sm" htmlFor="status-in-progress">
@@ -348,9 +312,9 @@ const EditActivityModal = ({
               <div className="flex items-center">
                 <Switch
                   id="status-paid"
-                  checked={activityStatusPaid}
+                  checked={statusPaid}
                   onCheckedChange={(checked) => {
-                    setActivityStatusPaid(checked)
+                    setStatusPaid(checked)
                   }}
                 />
                 <Label className="ml-2 text-sm" htmlFor="status-paid">
@@ -366,7 +330,7 @@ const EditActivityModal = ({
                   </Button>
                 </DialogClose>
                 <Button onClick={handleSubmit} type="button">
-                  Editar
+                  Adicionar
                 </Button>
               </div>
             </DialogFooter>
@@ -377,4 +341,4 @@ const EditActivityModal = ({
   )
 }
 
-export default EditActivityModal
+export default AddActivityModal
