+++ src/components/student-profile/activity/remove-activity-modal.tsx
